Update Find the Max & Min Value in an Inputted List.py
Prompts the User to input values.  Finds and prints the minimum and maximum value in the list.
Input 'done' to finish constructing list and run code.

+++ Find the Max & Min Value in an Inputted List.py
@@ -8,7 +8,7 @@ while True :
 	if inp == 'done' :
 		break
 	try:
-		fval = int(inp)
+		fval = float(inp)
 		numlist.append(fval)
 	except:
 		print('Invalid input')
